Unreviewed. Fix GTK+ unit tests crashing after r218922.
We need to check if icon database exists and is open now on dispose and API entry points.

* UIProcess/API/glib/WebKitFaviconDatabase.cpp:
(webkitFaviconDatabaseDispose):
(webkit_favicon_database_get_favicon_uri):
(webkit_favicon_database_clear):

Canonical link: https://commits.webkit.org/190799@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@218923 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/API/glib/WebKitFaviconDatabase.cpp
@@ -88,9 +88,8 @@ static void webkitFaviconDatabaseDispose(GObject* object)
 {
     WebKitFaviconDatabase* database = WEBKIT_FAVICON_DATABASE(object);
 
-    WebKitFaviconDatabasePrivate* priv = database->priv;
-    if (priv->iconDatabase->isOpen())
-        priv->iconDatabase->close();
+    if (webkitFaviconDatabaseIsOpen(database))
+        database->priv->iconDatabase->close();
 
     G_OBJECT_CLASS(webkit_favicon_database_parent_class)->dispose(object);
 }
@@ -458,6 +457,9 @@ gchar* webkit_favicon_database_get_favicon_uri(WebKitFaviconDatabase* database, 
     g_return_val_if_fail(pageURL, nullptr);
     ASSERT(RunLoop::isMain());
 
+    if (!webkitFaviconDatabaseIsOpen(database))
+        return nullptr;
+
     String iconURLForPageURL = database->priv->iconDatabase->synchronousIconURLForPageURL(String::fromUTF8(pageURL));
     if (iconURLForPageURL.isEmpty())
         return nullptr;
@@ -474,6 +476,9 @@ gchar* webkit_favicon_database_get_favicon_uri(WebKitFaviconDatabase* database, 
 void webkit_favicon_database_clear(WebKitFaviconDatabase* database)
 {
     g_return_if_fail(WEBKIT_IS_FAVICON_DATABASE(database));
+
+    if (!webkitFaviconDatabaseIsOpen(database))
+        return;
 
     database->priv->iconDatabase->removeAllIcons();
 }
